+++ libraries/Kelvin2RGB/Kelvin2RGB.h
@@ -2,52 +2,52 @@
 //
 //    FILE: Kelvin2RGB.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.7
+// VERSION: 0.2.0
 //    DATE: 2018-01-31
 // PURPOSE: Arduino library for converting temperature to RGB values
 //     URL: https://github.com/RobTillaart/Kelvin2RGB
 
 
-#define KELVIN2RGB_LIB_VERSION            (F("0.1.7"))
+#define KELVIN2RGB_LIB_VERSION            (F("0.2.0"))
 
 #include "Arduino.h"
 
 
-//
-//  Based upon article Tanner Helland  and Neil Bartlett
+//  Library based upon article Tanner Helland and Neil Bartlett
 //  http://www.tannerhelland.com/4435/convert-temperature-rgb-algorithm-code/
-//  http://www.zombieprototypes.com/?p=210
+//  https://en.wikipedia.org/wiki/Color_temperature#Categorizing_different_lighting
+
+
+
+//  Values in table below based on 
 //  https://en.wikipedia.org/wiki/Color_temperature#Categorizing_different_lighting
 //
+//             DAY LIGHT SETTING   = TEMPERATURE
 //
-//  based on https://en.wikipedia.org/wiki/Color_temperature#Categorizing_different_lighting
-//
-//      DAY LIGHT SETTING      TEMPERATURE
-//
-#define DLS_dark                     0
-#define DLS_match                 1700
-#define DLS_sodiumLamp            1700
-#define DLS_candleFlame           1850
-#define DLS_sunrise               1850
-#define DLS_sunset                1850
-#define DLS_bulb                  2400
-#define DLS_bulbSoftWhite         2550
-#define DLS_LEDlamp               2700
-#define DLS_warmWhite             3000
-#define DLS_studioLight           3200
-#define DLS_studioCPlight         3350
-#define DLS_daylightHorizon       5000
-#define DLS_flashLight            5700
-#define DLS_xenonLight            6200
-#define DLS_dayLightBright        6500
-#define DLS_normal                6500
-#define DLS_screenlow             6500
-#define DLS_screenMed             8000
-#define DLS_screenHigh            9500
-#define DLS_polewardSky0         15000
-#define DLS_polewardSky1         19000
-#define DLS_polewardSky2         23000
-#define DLS_polewardSky3         27000
+const uint16_t DLS_dark            =     0;
+const uint16_t DLS_match           =  1700;
+const uint16_t DLS_sodiumLamp      =  1700;
+const uint16_t DLS_candleFlame     =  1850;
+const uint16_t DLS_sunrise         =  1850;
+const uint16_t DLS_sunset          =  1850;
+const uint16_t DLS_bulb            =  2400;
+const uint16_t DLS_bulbSoftWhite   =  2550;
+const uint16_t DLS_LEDlamp         =  2700;
+const uint16_t DLS_warmWhite       =  3000;
+const uint16_t DLS_studioLight     =  3200;
+const uint16_t DLS_studioCPlight   =  3350;
+const uint16_t DLS_daylightHorizon =  5000;
+const uint16_t DLS_flashLight      =  5700;
+const uint16_t DLS_xenonLight      =  6200;
+const uint16_t DLS_dayLightBright  =  6500;
+const uint16_t DLS_normal          =  6500;
+const uint16_t DLS_screenLow       =  6500;
+const uint16_t DLS_screenMed       =  8000;
+const uint16_t DLS_screenHigh      =  9500;
+const uint16_t DLS_polewardSky0    = 15000;
+const uint16_t DLS_polewardSky1    = 19000;
+const uint16_t DLS_polewardSky2    = 23000;
+const uint16_t DLS_polewardSky3    = 27000;
 
 
 class Kelvin2RGB
@@ -58,14 +58,14 @@ public:
   //  resets all variables to zero.
   void begin();
   [[deprecated("Use begin() instead")]]
-  void reset() { begin(); };  // obsolete soon
+  void reset() { begin(); };  //  obsolete 0.3.0
 
   //  temperature = 0..65500   brightness = 0.0 .. 100.0%
   void convert_TH(float temperature, float brightness = 100.0);
   void convert_NB(float temperature, float brightness = 100.0);
 
   float    temperature();
-  float    brightness();;
+  float    brightness();
 
   //  returns 0.0 .. 1.0
   float    red();
@@ -79,10 +79,8 @@ public:
 
   uint32_t RGB();       //  32 bit colour  (only 3 bytes used)
   uint16_t RGB565();    //  16 bit colour  (2 bytes)
-
-  //  Experimental 0.1.3
-  uint32_t CMYK();
   uint32_t BGR();
+  uint32_t CMYK();
 
 
 private: